explainer: make it a Python project under tools/, with both video styles
Three changes that belong together.

Moved out of scripts/. That directory holds CI and legal glue; this takes a run
trace and produces explainer reports and shareable video, which is a tool. It
also stays out of the swath-* root namespace on purpose: every swath-* directory
at the root is a Gradle module listed in settings.gradle.kts, so a swath-explainer/
that was not one would invite `./gradlew :swath-explainer:build` and fail.

Split the single ~1800-line file into a package. The three HTML templates are now
real .html files rather than string literals, so the CSS and canvas code are
editable as themselves and render.py just substitutes one placeholder. Code splits
along the seams that were already there: keys (the mass-CDF projection), trace
(lenient reading), model (the reduction), render, cli, selftest. pyproject.toml
declares no runtime dependencies — deliberately, so the repo keeps its
zero-dependency property — and provides a swath-explainer command; explainer.py
stays as a shim so the tool runs from a clone with nothing set up.

Both video styles are kept and selectable via --video-style. `strip` (default) is
the algorithm as it happens: the keyspace cut into seed ranges, all of them
draining in parallel with each bar filling to its worker's cursor, then stealing
subdividing whatever is left. Seed cut points are drawn as fence posts, the
heaviest guesses labelled with their real bounds, and prefix families named over
the span their objects occupy. `map` is the earlier space-time carving diagram —
denser and closer to a flame graph, harder to narrate cold, but it shows
split-chain depth and which regions outlived the run.

README covers running it end to end on any bucket, the layout, the two axis and
colour decisions, determinism, and the trace's sensitivity class.
