+++ tools/explainer/capture-video.js
